requirements.txt
keyboard==0.13.5
pyautogui==0.9.54
Pillow==11.1.0
google-generativeai==0.3.2
